refactor: ShareButton과 ShareMenu 통합 및 ReactionBox -> Reaction 이름 변경
